♻️ Refactor MultiSelectDropDown (#640) * Move SVG icons out into their own components within `/Icons` * Replace onBlur/setTimeout hack with a click handler and a simplified blur handler
