+++ modules/components/src/Icons/CheckIcon.js
@@ -1,0 +1,16 @@
+import React from 'react';
+
+const CheckIcon = () => {
+  return (
+    <svg width={9} height={9} viewBox="0 0 8 6">
+      <g fill="currentColor">
+        <path
+          d="M.225 4.877c-.687-.686.357-1.73 1.043-1.044L2.596 5.16l4.136-4.135c.686-.687 1.73.357 1.043 1.044L3.118 6.726c-.289.289-.756.289-1.044 0L.224 4.877z"
+          transform="translate(-1105 -366) translate(0 63) translate(265 17) translate(0 237) translate(830 11) translate(0 27) translate(1 6) translate(9 4)"
+        />
+      </g>
+    </svg>
+  );
+};
+
+export default CheckIcon;
