+++ modules/components/src/Icons/index.js
@@ -1,0 +1,3 @@
+export { default as ArrowIcon } from './ArrowIcon';
+export { default as CheckIcon } from './CheckIcon';
+export { default as ResetIcon } from './ResetIcon';
